backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.7.3
pydantic-settings==2.1.0

# memU (mem0)
mem0ai==0.1.0

# Database
qdrant-client==1.9.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
alembic==1.13.1

# AI
openai==1.33.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
